Refactor latest selector tests out of chat hotspot
